requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
django-unfold==0.90.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.2
pluggy==1.6.0
pycparser==3.0
Pygments==2.20.0
PyJWT==2.11.0
pytest==9.0.3
pytest-django==4.12.0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.12.0